Refactored Header component to use polymorphic Button component and moved RomDirectoryPage into pages directory
