--- pokemon-collection/src/components/card/helpers/CardDescription.tsx
@@ -1,8 +1,0 @@
-// types
-import type CardType from '../../../types/Card';
-
-const CardDescription = ({ description }: CardType) => {
-  return <p>{description}</p>;
-};
-
-export default CardDescription;
