--- pokemon-collection/src/components/card/helpers/CardImage.tsx
@@ -1,14 +1,0 @@
-// types
-import type CardType from '../../../types/Card';
-
-const CardImage = ({ img, title }: CardType) => {
-  return (
-    <img
-      className="w-full h-full object-cover rounded-[15px]"
-      src={img}
-      alt={title}
-    />
-  );
-};
-
-export default CardImage;
