+++ pokemon-collection/src/layouts/header/Header.tsx
@@ -1,47 +1,54 @@
+// images
 import dean from '../../assets/dean_github.png';
 import zach from '../../assets/zach_github.png';
 import efe from '../../assets/efe_github.png';
+// components
+import Button from '../../components/ui/Button';
 
-interface HeaderProp {
+interface HeaderProps {
   title?: string;
 }
 
-function Header({ title = 'Pokemon ROM Hack Collection' }: HeaderProp) {
+const Header: React.FC<HeaderProps> = ({
+  title = 'Pokemon ROM Hack Collection',
+}): React.JSX.Element => {
   return (
     <header className="flex justify-between items-center">
       <h1 className="title">{title}</h1>
 
       <ul className="list-none flex p-0">
         <li className="pt-[15px] pr-[10px]">
-          <a href="https://github.com/dmartin34-rrc">
+          <Button href="https://github.com/dmartin34-rrc">
             <img
+              className="rounded-full w-[50px]"
               src={dean}
               alt="Dean's GitHub Profile Picture"
-              className="rounded-full w-[50px]"
             />
-          </a>
+          </Button>
         </li>
+
         <li className="pt-[15px] pr-[10px]">
-          <a href="https://github.com/zachary-lam">
+          <Button href="https://github.com/zachary-lam">
             <img
+              className="rounded-full w-[50px]"
               src={zach}
               alt="Zach's GitHub Profile Picture"
-              className="rounded-full w-[50px]"
             />
-          </a>
+          </Button>
         </li>
+
         <li className="pt-[15px] pr-[10px]">
-          <a href="https://github.com/EfeOnota">
+          <Button href="https://github.com/EfeOnota">
             <img
+              className="rounded-full w-[50px]"
               src={efe}
               alt="Efe's GitHub Profile Picture"
-              className="rounded-full w-[50px]"
             />
-          </a>
+          </Button>
         </li>
       </ul>
     </header>
   );
-}
+};
 
 export default Header;
